TSMM/requirements.txt
absl-py==2.1.0
accelerate==1.0.1
annotated-types==0.7.0
anyio==4.4.0
arviz==0.22.0
asn1crypto==1.5.1
asttokens==2.4.1
astunparse==1.6.3
attrs==25.3.0
backcall==0.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
boto3==1.39.8
botocore==1.39.8
cachetools==5.3.3
cattrs==25.1.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cmdstanpy==1.2.5
colorama==0.4.6
contourpy==1.3.1
cryptography==42.0.8
cycler==0.12.1
dash==3.0.1
dash-bootstrap-components==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.6.1
docker==7.1.0
docopt==0.6.2
ecdsa==0.19.0
email_validator==2.1.1
ephem==4.2
et_xmlfile==2.0.0
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.19.1
filelock==3.16.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.56.0
fpdf==1.7.2
fsspec==2024.10.0
fuzzywuzzy==0.18.0
gast==0.5.4
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
h5netcdf==1.6.3
h5py==3.11.0
holidays==0.76
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface==0.0.1
huggingface-hub==0.26.1
idna==3.7
importlib_metadata==8.5.0
importlib_resources==6.5.2
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
keras==2.15.0
kiwisolver==1.4.8
libclang==18.1.1
lxml==6.0.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.6
mdurl==0.1.2
middleware==1.2.3
mistune==3.0.2
ml-dtypes==0.2.0
mpmath==1.3.0
nbclient==0.9.0
nbconvert==7.16.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
orbit==0.2
orbit-ml==1.1.4.9
orjson==3.10.4
packaging==24.1
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
passlib==1.7.4
patsy==1.0.1
pickleshare==0.7.5
pillow==11.1.0
pipreqs==0.5.0
platformdirs==4.2.0
plotly==5.24.1
prompt-toolkit==3.0.43
prophet==1.1.7
protobuf==4.25.3
psutil==6.1.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pyephem==9.99
Pygments==2.18.0
PyJWT==2.10.1
pyngrok==7.2.3
pyOpenSSL==25.1.0
pyparsing==3.2.3
pyro-api==0.1.2
pyro-ppl==1.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
redis==5.0.5
referencing==0.33.0
regex==2024.9.11
requests==2.32.3
requests-cache==1.2.1
requests-oauthlib==2.0.0
retrying==1.3.4
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
s3transfer==0.13.0
safetensors==0.4.5
scikit-learn==1.4.1.post1
scipy==1.13.1
seaborn==0.13.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
snowflake==1.6.0
snowflake-connector-python==3.16.0
snowflake._legacy==1.0.1
snowflake.core==1.6.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.30
stack-data==0.6.3
stanio==0.5.1
starlette==0.37.2
statsmodels==0.14.5
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
tinycss2==1.2.1
tokenizers==0.20.1
tomlkit==0.13.3
torch==2.5.0
tornado==6.4
tqdm==4.66.5
traitlets==5.14.1
transformers==4.46.0
typer==0.12.3
types-cffi==1.16.0.20240331
types-pyOpenSSL==24.1.0.20240425
types-redis==4.6.0.20240425
types-setuptools==70.0.0.20240524
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
url-normalize==2.2.1
urllib3==2.2.1
uvicorn==0.30.1
watchfiles==0.22.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.3
wrapt==1.14.1
xarray==2025.7.0
xarray-einstats==0.9.1
xgboost==3.0.2
yarg==0.1.9
zipp==3.21.0